+++ src/main.js
@@ -1,8 +1,10 @@
 import kaboom from "kaboom";
-import "kaboom/global";
+import { layerPlugin } from "kaboom-extra";
+import { GameData } from "./classes/gameData";
 import { loadAssets } from "./loader";
-import { gameScene } from "./scene_game";
-import { sceneSongSelection } from "./scene_song_selection";
+import { loadGameScene } from "./scenes/scene_game";
+import { sceneSongSelection } from "./scenes/scene_song_selection";
+import { loadClickScene } from "./scenes/scene_click";
 
 export const k = kaboom({
     width: 720,
@@ -12,12 +14,22 @@ export const k = kaboom({
     texFilter: "nearest",
     crisp: true,
     font: "happy",
+    background: [0, 0, 0],
+    plugins: [layerPlugin],
 });
 
-export const gameData = {
-    "songs": [],
-}
+export const gameData = new GameData();
+
+k.layers([
+    "background",
+    "note",
+    "player",
+    "sword",
+    "default",
+    "ui",
+], "default");
 
 loadAssets();
-gameScene();
+loadGameScene();
 sceneSongSelection();
+loadClickScene();