+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Alpha 1.2.0 - TBD
+## Mechanics
+- New note type: "Slider"
+
+## Graphics
+- New slider sprite: "Green Apple"
+- New song in main menu desing
+
 # Alpha 1.1.0 - 12/11/2023
 ## Mechanics
 - Combo